Started Updating Lang The lang system should be updated to use Yaml eventually. I've started by moving all the variables stored in Lang.java into an Enum MessageTypes.
